requirements-dev.txt
black==22.10.0
Django==4.1.2
django-filter==22.1
flake8==5.0.4
isort==5.10.1
Pillow==9.2.0
six==1.16.0
django-environ==0.9.0
whitenoise==6.2.0
coverage==6.5.0
